Libraries to support using go-git
* This PR contains code that will enable using git programmatically via
  go-git as opposed to shelling out to the git CLI

This PR updates RepoHelper to support cloning, branching, committing, and pushing
 code using go-git

 * In a follow on PR we will update Syncer to use this functionality rather than
   shelling out to git.

 * Create some integration tests for testing this. This is behind a go build flag.

 * This leads to the function signature of CreatePR and PullRequestForBranch being changed
  since some of the function args should not be obtained from the struct rather than passed
  as command line arguments.

Add a GoGit Auth implementation that works for authenticating as a GitHub App

* This is just basic auth with the username x-access-code
* We also need to generate the token which is used as the password on
  every call since the token expires.

Related to https://github.com/jlewi/hydros/issues/3 ; this is tracking the migration
from using git CLI to using go-git.

Fix some but not all deprecated reefrences to ioutil.

Update the README to describe some of the other code in this repository