Improved handling of PR merging (#22)
* This PR reduces the reliance on GitHub's automerge feature by adding
improved handling of PR merging in the syncer.

* If a PR exists the syncer will attempt to merge it and wait for it to
merge this allows the syncer to push any changes that might be blocked
on the pending PR

* After creating a new PR with the latest changes the sycner will wait
for the PR to be merged. This provides better handling in the case of
dev takeover when automerge isn't (or can't be enabled) on the
repository. There appears to be a bit of a race condition where the PR
may not be merged right away even thought it should be immediately
mergeable but this PR fixes that by having a retry loop that will wait
for the PR to me merged and periodically retry merging

* Fix #21
* Fix #17