Create a repo controller (#59)
A repo controller will be used to periodically sync all the resources
defined in a repository.

This may turn out to be only an interim solution once we have K8s
controllers for the different resources. Or it may end up being used by
the CLI to do manual deployments.